+++ test/test_queue.js
@@ -310,6 +310,11 @@ describe('Queue', () => {
           expect(job.id).to.be.ok;
           expect(job.data.foo).to.be.eql('bar');
         })
+        .then(
+          queueFoo.bclient.client('GETNAME').then(name => {
+            expect(name).to.be.eql(queueFoo.clientName());
+          })
+        )
         .then(() => {
           return queueQux.add({ qux: 'baz' }).then(job => {
             expect(job.id).to.be.ok;
@@ -2262,6 +2267,48 @@ describe('Queue', () => {
             attempts: 3,
             backoff: {
               type: 'custom'
+            }
+          }
+        );
+      });
+      queue.on('completed', () => {
+        const elapse = Date.now() - start;
+        expect(elapse).to.be.greaterThan(3000);
+        done();
+      });
+    });
+
+    it('should pass strategy options to custom backoff', function(done) {
+      this.timeout(12000);
+      queue = utils.buildQueue('test retries and backoffs', {
+        settings: {
+          backoffStrategies: {
+            custom(attemptsMade, err, strategyOptions) {
+              expect(strategyOptions.id).to.be.equal('FOO42');
+              return attemptsMade * 1000;
+            }
+          }
+        }
+      });
+      let start;
+      queue.isReady().then(() => {
+        queue.process((job, jobDone) => {
+          if (job.attemptsMade < 2) {
+            throw new Error('Not yet!');
+          }
+          jobDone();
+        });
+
+        start = Date.now();
+        queue.add(
+          { foo: 'bar' },
+          {
+            attempts: 3,
+            backoff: {
+              type: 'custom',
+              options: {
+                id: 'FOO42'
+              }
             }
           }
         );